Fix putting ARG_SCOPE_NAME to arguments if arguments is null
If Fragment will be created without arguments, then arguments will be null and arguments?.putString()  would have no effect. In this case it's not critical, but it's potentially can be a cause of bugs in the future, for example if somebody will copy this code to another fragment which will be created without arguments.